[xen] Add domctl getdomaininfo and log a particular error if it fails with EACCES
Only getdomaininfo can be called from a stubdom and only then with a suitable
XSM policy loaded. If we get EACCES we mention XSM on the console...

Note: this isn't part of the xen guest ABI but rather is only used
by the xen tools, which are always matched with the hypervisor.
This means that this code would need compiling for each xen version.
-- perhaps it should actually be moved directly into the xenstore
   daemon code.

Note: the only definition of the xen_domctl_getdomaininfo struct is
in the C header file and there is nothing stopping it from evolving
in arbitrary ways. Rather than duplicate the definition into a cstruct
decl, we drop down to C for 'sizeof' and 'parse'.

Note: we only bother parsing the exact fields needed by the xenstore
daemon code -- another argument for moving the code there.